Fix repository struct defaults for varchar | text

+++ common/repositories/instance_list_repository.h
@@ -27,13 +27,13 @@
 class InstanceListRepository {
 public:
 	struct InstanceList {
-		int id;
-		int zone;
-		int version;
-		int is_global;
-		int start_time;
-		int duration;
-		int never_expires;
+		int  id;
+		int  zone;
+		int8 version;
+		int8 is_global;
+		int  start_time;
+		int  duration;
+		int8 never_expires;
 	};
 
 	static std::string PrimaryKey()
